+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/posting/api/controller/PostController.java
@@ -1,6 +1,6 @@
 package com.ambiguous.buyornot.posting.api.controller;
 
-import com.ambiguous.buyornot.chatting.api.support.response.ApiResult;
+import com.ambiguous.buyornot.common.support.response.ApiResult;
 import com.ambiguous.buyornot.posting.api.controller.request.CreatePostDto;
 import com.ambiguous.buyornot.posting.api.domain.PostService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,8 +21,8 @@ public class PostController {
             @RequestBody CreatePostDto dto
     ) {
         Long userId = 1L;
-        String userNickname = "nickName";
-        postService.createPost(stockId, userId, userNickname, dto);
+        String nickname = "nickName";
+        postService.createPost(stockId, userId, nickname, dto);
 
         return ApiResult.success("게시물이 생성되었습니다.");
     }
