+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/user/api/controller/request/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.ambiguous.buyornot.user.api.controller.request;
+
+public class LoginRequest {
+}
