+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/user/api/controller/request/RefreshTokenRequest.java
@@ -1,0 +1,4 @@
+package com.ambiguous.buyornot.user.api.controller.request;
+
+public class RefreshTokenRequest {
+}
