+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/user/api/domain/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.ambiguous.buyornot.user.entity;
+package com.ambiguous.buyornot.user.api.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
