+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/user/api/domain/UserRole.java
@@ -1,0 +1,6 @@
+package com.ambiguous.buyornot.user.api.domain;
+
+public enum UserRole {
+    USER,
+    ADMIN
+}
