--- BuyOrNot/src/main/java/com/ambiguous/buyornot/user/dto/request/LoginRequest.java
@@ -1,4 +1,0 @@
-package com.ambiguous.buyornot.user.dto.request;
-
-public class LoginRequest {
-}
