--- BuyOrNot/src/main/java/com/ambiguous/buyornot/user/dto/request/RefreshTokenRequest.java
@@ -1,4 +1,0 @@
-package com.ambiguous.buyornot.user.dto.request;
-
-public class RefreshTokenRequest {
-}
