--- BuyOrNot/src/main/java/com/ambiguous/buyornot/user/entity/UserRole.java
@@ -1,6 +1,0 @@
-package com.ambiguous.buyornot.user.entity;
-
-public enum UserRole {
-    USER,
-    ADMIN
-}
